fix: 카프카 로직 추가

+++ customercenter/src/main/java/elearning/infra/MyPageRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 @RepositoryRestResource(collectionResourceRel = "myPages", path = "myPages")
 public interface MyPageRepository
     extends PagingAndSortingRepository<MyPage, Long> {
-    List<MyPage> findByEnrollmentId(Integer enrollmentId);
+    List<MyPage> findByEnrollmentId(Long enrollmentId);
 }
